Change types of damage/hp related props to int

+++ include/game-models/AbilityEndowedPlayer/ability-endowed-player.cpp
@@ -22,7 +22,7 @@ AbilityEndowedPlayer::AbilityEndowedPlayer(
 	PlayerTeamId teamId,
 	int initialHitPoints,
 	int initialAmmo,
-	double damage,
+	int damage,
 	long long abilityCoolDown_ms
 ) :
 	Player(
